changed to depend on animation4j 0.0.14-SNAPSHOT, removed delta parameter from Screen and GameState update and render methods and added a setDelta(float delta) method called by the container to set current delta time, added proper methods getDelta and getDeltaInMs to GameState to be used by its sub classes

+++ commons-gdx-core/src/test/java/com/gemserk/commons/gdx/GameStateTransitionTest.java
@@ -40,7 +40,7 @@ public class GameStateTransitionTest {
 		}
 
 		@Override
-		public void set(GameState t, int time) {
+		public void set(GameState t, float time) {
 			if (current == null || time <= 0) {
 				current = t;
 				started = true;
